Migrate FrameworkTests to addSourceLines
Replace the 8 remaining addSourceFile calls in FrameworkTests.java with
addSourceLines and inlined text blocks, delete the 8 testdata files that
are no longer referenced, and drop the class-level
@SuppressWarnings("deprecation").

License headers are stripped from the inlined sources, matching the
approach taken in #1517.

Partial progress on #1425.
